feat(sitemap): generate with Next.js
