+++ .vscode/settings.json
@@ -13,25 +13,25 @@
   "prettier.configPath": ".prettierrc.json",
   "workbench.colorCustomizations": {
     "activityBar.activeBorder": "#06b9a5",
-    "activityBar.activeBackground": "#fbed80",
-    "activityBar.background": "#fbed80",
+    "activityBar.activeBackground": "#f7b13c",
+    "activityBar.background": "#f7b13c",
     "activityBar.foreground": "#15202b",
     "activityBar.inactiveForeground": "#15202b99",
-    "activityBarBadge.background": "#06b9a5",
-    "activityBarBadge.foreground": "#15202b",
+    "activityBarBadge.background": "#057a4f",
+    "activityBarBadge.foreground": "#e7e7e7",
     "commandCenter.border": "#15202b99",
-    "sash.hoverBorder": "#fbed80",
-    "statusBar.background": "#f9e64f",
+    "sash.hoverBorder": "#f7b13c",
+    "statusBar.background": "#f59e0b",
     "statusBar.foreground": "#15202b",
-    "statusBarItem.hoverBackground": "#f7df1e",
-    "statusBarItem.remoteBackground": "#f9e64f",
+    "statusBarItem.hoverBackground": "#c57f08",
+    "statusBarItem.remoteBackground": "#f59e0b",
     "statusBarItem.remoteForeground": "#15202b",
-    "titleBar.activeBackground": "#f9e64f",
+    "titleBar.activeBackground": "#f59e0b",
     "titleBar.activeForeground": "#15202b",
-    "titleBar.inactiveBackground": "#f9e64f99",
+    "titleBar.inactiveBackground": "#f59e0b99",
     "titleBar.inactiveForeground": "#15202b99"
   },
-  "peacock.color": "#f9e64f",
+  "peacock.color": "F59E0B",
   "typescript.tsdk": "node_modules/typescript/lib",
   "typescript.enablePromptUseWorkspaceTsdk": true
 }
