+++ package.json
@@ -4,7 +4,6 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "postbuild": "next-sitemap --config next-sitemap.js",
     "start": "next start",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx --ignore-path .gitignore .",
     "format": "prettier --write \"./**/*.{js,jsx,ts,tsx,css,md,json}\" --config ./.prettierrc.json --ignore-path ./.gitignore"
@@ -31,7 +30,6 @@
     "gray-matter": "^4.0.3",
     "gsap": "^3.11.5",
     "next-mdx-remote": "^4.4.1",
-    "next-sitemap": "^4.0.7",
     "postcss": "^8.4.23",
     "prettier": "^2.8.8",
     "react-lottie-player": "^1.5.4",
